FinanceHooks/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
flask

uuid_v7
uuid7
azure.cosmos
azure.storage.blob 
azure.identity
azure-ai-vision-imageanalysis